Remove excess code from slime_helper
There's tab completion for screen sessions in slime.vim, so most of this
isn't necessary. Use the existing code and just autoload the screen
session if there's only one.

Now, this plugin is mostly just to fix the mapping so that \r on first
invocation doesn't give you `` at the session prompt.