Add function to simplify git merge cleanup
git clean isn't useful because there are some files that are not
ignored, but are also not in repo (maybe I should do something about
that).

Instead, add a function to allow you to select lines from the git
commit, find their absolute path and then delete them (will attempt to
move to trash before deleting).