Simplify ctrlp maps: use a common leader
My less-common ctrlp maps are prefixed with C-o
C-S-o is actually the same as C-o, so this just swaps out the buffer
select (which I don't use since MRU is so easy).

The more common open file (A-S-o) and MRU (A-S-m) get their own maps
since I use them so often.

Also move the extensions list down by the maps so I keep them in sync.