listchars tab char; change function to command
I can't see u9656 on Windows (I set this up initially on Ubuntu). Change
listchars tab setting to use a char I can see. (Hopefully it works when
I switch back to Ubuntu.)

Include the char for eol. This one seems to be really common, but I
don't like having so many chars on the screen.

Instead of calling ListCharToggle, make it a function so it's a little
easier to type.