Rename vim-scratch to vim-aside
!!!
This requires you to rename your ~/.vim-scratch file to ~/.vim-aside
!!!

Seems that the term 'scratch' commonly refers to a buffer _without_ a
backing file, so vim-scratch is kind of the opposite.

Also fix persistent_undo check for vimscratch files. I had previously
used *.scratch which is actually a visual programming language.

Also added a map to quickly open ~/.todo.org that's similar to
vim-aside: Alt-F1. Hopefully that works on all platforms.