Simplify loading of system plugins
Some tools offer vim integration and provide their own vimscript.
Instead of copying it to ~/.vim, use david#load_system_vimscript to
source it. The function returns whether sourcing was successful.

This function should be guarded with loaded_NAME so the plugins can be
disabled in local.vim (when that package is not and will not be
installed).

The function's return value is a good value for the loaded_NAME guard
(so it is defined but zero when we failed to load).

Use this function to load some python integration and add the plugins to
local_templates.