Fix unintentional permission changes on Windows
Many of my git commits show file permissions changed from 644 to 755. I
don't want this. It appears to be caused by the way vim writes files: It
uses backupcopy=auto by default which seems to do take the "no" path.
According to the docs, it renames the file and writes a new one -- so it
moves the edited files away and creates a new file with your modified
contents. For some reason, it doesn't inherit the file permissions
(probably because they don't normally exist in Windows -- they're only
visible through cygwin).

Using backupcopy=yes fixes this. It also apparently fixes overwriting of
symlinks.