Disable tearoff menus and fancy tabline (in win32)
I don't use tearoff menus anymore, so disable them (I only have the
menubar for when other people come to my desk).

Disable the fancy tabline in Windows. It causes vim to re-position
itself when it creates the tabline (because it's using the gui system's
tabs). Without this option, it uses ascii tabs which look bad, but I
don't use tabs often so I can deal.

Source for tabline fix:
http://stackoverflow.com/questions/13251522/why-does-gvim-resize-and-reposition-itself-after-some-actions