Remove dosbatch.vim because it's shipped with vim
I added this back in 2010 because it was newer than whatever vim shipped
with at the time. Vim 7.3 and 7.4 both ship with this version of syntax,
so don't duplicate it.

I also modified it to remove syntax highlight for one case of
dosbatchLabel -- where the word starts with a label because I was
working somewhere that used :call a lot. I don't anymore and I think
that usage is incorrect, so I'm okay with losing that change. (Also
since the built-in syntax file was loaded, my change had no effect.)