Replace cinkeys with cinoptions for #
I don't need to set cinkeys if I'm not using cindent. I probably
want # to always be properly indented. Evaluating indent when I press #
is not the problem -- it's setting the indentation incorrectly.

cinoptions=#1 makes vim try to figure out if the # is a comment. It
seems to work, so let's use that. In cases where that fails (and I'm
using cindent), I could do an ftplugin:

    setlocal cinkeys-=0#