Replace yankstack with unite-history_yank-cycle
yankstack maps yank/delete keys so it can keep a yank history, but this
breaks some macro recording and makes me not trust some basic vim
commands.

unite's history_yank uses autocmds to collect recent yanks. While it
might not be perfect, it doesn't map keys. I want to use it as a back
end for a yankstack-like interface. Since all I use from yankstack are
the previous/next maps, that's all that's provided.

unite-history_yank-cycle would be better if it could use
UnitePrevious/Next or used unite's internal yank list, but that's for
another day.