Use <Leader>w for window management
Apply this map:
    nnoremap <Leader>w <C-w>

That also requires some changes in how I was mapping <Leader>w. I'm
going to try to be consistent with <C-w>, but I've got to lose some old
habits.

I've moved the QFEdit map to only apply in the quickfix buffer. I don't
use it elsewhere. I don't have a good map for it so wD will have to do.