Add a basic task plugin
Opens a specific plan file with orgmode filetype. Mostly this is just an
interface for opening a prioritized task list and relies on orgmode for
all the real work.

Change my filenames. My old todo is huge and not a useful todo. Rename
that to backlog and make the new thing plan.