Better support Bash on Windows for initial setup
Bash on Windows is close to cygwin, but there are some differences. The
major one solved here is where drives are linked from: /cygdrive vs
/mnt.

I wasn't using cygdrive before, but I think it makes sense to link out
of a directory instead of my funky stuff. Update to support this new
workflow.