Implement gf for vim
Jump to file that may have defined vim symbol.

A bit weird because it's not looking at filenames (vim doesn't have
includes), but it's not a tag: we're guessing based on filename like
normal gf behaviour (not looking at an index).

Guesses the file that defined the input symbol and use Vopen
(vim-scriptease) to open it.