Make svn augroups unique per buffer
Since I'm adding this autocmd to each buffer, I can't clear a common
augroup. Instead, make one per buffer and clear it.

We reach this code on the first time and after the autocmd executed, so
we still want to clear the autocmd. Could clear after execution, but
this seems simpler.