Only use syntax folding when available
Default to indent.
If we have fastfold and the syntax supports it, use syntax folding.

This requires us to build a list of filetypes that support syntax
folding. We naively search through our filetypes for something that
looks like a folding definition.

We could extend this setup to setting autocmds to set foldmethod for
each filetype, but this approach feels more controlled (we can determine
when foldmethod is modified).

Test
validate that if json.vim sets foldmethod=marker that setting is not
clobbered by try_use_syntax_folds.