Don't change notgrep mode if project exists
Fix slow grep on work projects.

Only change grep mode when first setting up a project (so if we already
configured it, we don't stomp it with a worse configuration).

If I setup notgrep to work on a sub directory or with cindex, I don't
want to switch to the slower (and more naieve
NotGrepUseGrepRecursiveFrom).

(My grep now completes in 3 seconds instead of 500 seconds.)