Use shebang to determine py2 vs py3
Instead of only relying on pyxversion, look at the shebang. This allows
me to edit and run py2 or 3 files on the same system.

Setting the makeprg correctly here is important to make
PythonSetEntrypoint work correctly.

Test :w a new python file with python3 and python in shebang and it's
correctly detected. Change the shebang and S-space to reload and it
correctly updates makeprg.