Add FindUnusedModules
Find lua modules that aren't `require`d anywhere. Great for cleaning up
love2d superprojects before publishing to limit to the actual project.

Not exposed as a command or anything. Would be nice if I could customize
the "import" method, but I don't have any other uses for this at the
moment.