Add script that moves docs from README to code
I don't like readme-only documentation. The readme usually has other
stuff and I want to be able to jump to the function I'm calling and see
the docs (or see them in completion hints). So it's better to put the
documentation into the code and generate a page with documentation.

This script tries to generate LDoc-compatible comments (colon style).