Add Gpeekcommand for blame to show commit message
When viewing blame, I often want a peek at the commit message instead of
opening it -- especially because opening it opens at the current line
but I just wanted an idea of why this commit was made.

Now I can pause on a line or hit x to show the commit message.