Prevent autochdir into .git or fugitive
Try to fix "Failed to remove .git\rebase-merge" error after rebase.

This happened again and I checked process explorer and it was gvim.exe
that was holding on to the handle. My guess is that editing the rebase
todo list made me chdir into rebase-merge/ and then git couldn't delete
it.

I don't think I ever want to chdir for a fugitive buffer (most are not
files or are temp, so this is rare anyway). I also don't want to chdir
inside of .git to ensure this issue doesn't occur.

No testing to see if it's fixed, but here's hoping.