Fix gJ to use legacy behaviour with 1gJ
You can't pass a zero count. v:count is zero when no count was given. I
want to remove spaces in that case.

Use wants_legacy to capture the count before we issue any commands and
make what I'm checking clear and consistent.

Use gv in multi to ensure we apply gJ to the selection.