Remove tags file to ensure ctags rebuilds it
Fix "ctags: "tags" doesn't look like a tag file; I refuse to overwrite it."

Sometimes tag file gets corrupted (many @) written to the start and
ctags refuses to modify it. Not sure why things are going wrong, but
this should resolve.

My guess is that it has something to do with my continuous tag build.