mark: Use maximum palette for gui
I was using the original palette and the first colour from it looks like
my IncSearch color. I could also reverse that palette but it puts less
nice colours first. Could remove the first colour, but there aren't many
colours.

The first colour of this palette is still similar to IncSearch, but I'll
try it like this for a bit.

Using this palette also prevents my marks from being lost when the
colour cycles. I thought that was a limitation of the plugin, not the
palette!

Needs to go in vimrc because gvimrc is loaded *after* the plugin and it
only looks at these values on load.