nvim: Show file path in title
This defaults to off in nvim, but on in gvim. It's useful enough and my
terminals are short lived enough that I'd prefer to have it on than lose
it.

After not having this for a couple weeks, I've realized how much I use
it to know which file I'm editing. My lightline setup only shows full
path for inactive buffers since the title show the active buffer, so I'm
lost without it.