nvim: Don't lose place in file on reload
Fix cursor jumps to top of file when using Shift-Space to resolve link.

I use Shift-Space a lot to force lsp reload as well as resolve links.
Since nvim requires wiping the buffer and that clears marks and options,
do it sparingly. Also, even if we wiped, restore the window position
since we're loading up a different path with the same file contents.