Update :Shell to use new FALLBACK in zepl
I added a fallback config to zepl so I don't need to force the shell
anymore (which prevents using Shell to jump back to the repl).

I *could* use Repl to return and Shell to start, but it's easier to use
a single command for both. I use zepl more as a nicer wrapper around
:terminal than as an actual repl -- Ripple is usually better for that.