Change :Shell to always open the shell
Use the b:repl_config to force shell to open. Since zepl doesn't tell
you if it's already open, I don't know if I can pass cmd to open the
shell.

Remove the bang version. If I want a filetype repl, I should use :Repl.

Maybe zepl should open the repl if the cmd passed is the same as the one
that opened the repl? Seems like that'd be nice to identify multiple
repls.