nvim: Setup dap for debugging
Installed several plugins in daveconfig:

* mfussenegger/nvim-dap
* jay-babu/mason-nvim-dap.nvim
* rcarriga/nvim-dap-ui
* nvim-neotest/nvim-nio (dap-ui dependency)

This is configuration for them.

Tried to get it working for godot. (see upcoming vim-david change). I
can launch the game, see log, but my breakpoints don't stop the game. I
can see watch values when I use the editor's debugger to stop the game,
but can't trigger a break from within vim. : (

I saw somewhere to pass `launch_scene = true` to
dap.configurations.gdscript, but it's not listed in the args on the
vscode plugin[1] which seems to be the best docs for Godot DAP.

Copied Tom's config for gdb+cpp, but haven't tried it.

[1]: https://github.com/godotengine/godot-vscode-plugin?tab=readme-ov-file#gdscript-debugger