nvim: Use shorter filenames for quicker
Since it aligns all to the same column, it needs shorter filenames so
most lines show lots of content.

I wonder if removing the path for most filenames would be more useful...
I do notice that files are in different directories and some have the
same filename, so maybe not great. Shortening the path like tabs could
be good.