nvim: Add show_virtual_text_only_for_current_line
Turn off virtual_lines because it inserts fake newlines and moves code
around too much while I'm trying to edit. Seems like a nice idea, but
too bad it doesn't just draw over other lines.

Take this snippet of code from an issue and apply it:
https://github.com/neovim/neovim/issues/23526#issuecomment-1539580310

It seems to do a good job of actually only showing virtual_text for the
current line.

Add a keymap to revive virtual_lines when I really want to see a lot of
complaints.